removed the last committed trick
-user gets easily confused

+++ Xcode Project/Skedify/Skedify/ServerConnection.h
@@ -23,7 +23,7 @@
 
 + (ServerConnection *) sharedServerConnection;
 - (NSArray *) GetGroupList;
-- (NSArray *) GetGroupContacts: (int) groupIdentifier;
+- (NSArray *) GetGroupContacts: (NSInteger) groupIdentifier;
 - (void) addGroup:(Group *) theGroup;
 - (Group *) GetGroup:(NSInteger) index;
 @property (nonatomic, readonly) NSMutableArray *groupsList;
